1.完成商家商品服务模块 1）商品分类名称映射模板ID 2）模板ID映射品牌和规格数据 的缓存
2.完成商品搜索前台模块的
1）商品分类列表显示
2）显示第一个商品分类对应的品牌和规格列表数据

+++ pinyougou-sellergoods-service/src/main/java/com/lsf/pinyougou/sellergoods/service/impl/ItemCatServiceImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.Objects;
 
 import com.github.pagehelper.PageInfo;
+import com.lsf.pinyougou.pojo.TbItem;
 import com.lsf.pinyougou.pojogroup.TbItemCats;
 import com.lsf.pinyougou.sellergoods.service.ItemCatService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,7 +13,10 @@ import com.alibaba.dubbo.config.annotation.Service;
 import com.github.pagehelper.PageHelper;
 import com.lsf.pinyougou.dao.TbItemCatDao;
 import com.lsf.pinyougou.pojo.TbItemCat;
+import org.springframework.data.redis.core.RedisTemplate;
 import vo.PageResult;
+
+import javax.annotation.PostConstruct;
 
 
 /**
@@ -28,6 +32,20 @@ public class ItemCatServiceImpl implements ItemCatService {
      */
     @Autowired
     private TbItemCatDao tbItemCatDao;
+
+    @Autowired
+    private RedisTemplate redisTemplate;
+
+    @PostConstruct
+    private void init() {
+        // 当对象创建时，提前往 redis 中缓存 k=分类名称，v=模板ID
+        if (tbItemCatDao != null && redisTemplate != null) {
+            List<TbItemCat> list = tbItemCatDao.queryAll(new TbItemCat());
+            for (TbItemCat t : list) {
+                redisTemplate.boundHashOps("catNameToTypeID").put(t.getName(), t.getTypeId());
+            }
+        }
+    }
 
 
     /**
@@ -64,20 +82,48 @@ public class ItemCatServiceImpl implements ItemCatService {
 
 
     /**
-     * 添加商品分类
+     * 添加商品分类（数据库中添加，缓存中也添加）
      */
     @Override
     public void add(TbItemCat itemCat) {
+        // 往数据库中添加商品分类
         tbItemCatDao.insert(itemCat);
+
+        // 往 catNameToTypeID map 缓存中添加 k=分类名称，v=模板ID
+        redisTemplate.boundHashOps("catNameToTypeID").put(itemCat.getName(), itemCat.getTypeId());
     }
 
 
     /**
-     * 修改
+     * 修改（修改数据库中的内容，删除缓存中旧的商品分类名称）
      */
     @Override
     public void update(TbItemCat itemCat) {
-        tbItemCatDao.update(itemCat);
+
+        // 查找数据库中旧的商品分类的名称
+        TbItemCat tbItemCat = new TbItemCat();
+        tbItemCat.setId(itemCat.getId());
+        List<TbItemCat> result = tbItemCatDao.queryAll(tbItemCat);
+
+        if (result != null && result.size() > 0) {
+            // 修改数据库中的商品分类
+            tbItemCatDao.update(itemCat);
+
+            // 获取旧的商品分类名称
+            String oldName = result.get(0).getName();
+
+            if (!oldName.equals(itemCat.getName())) {
+                // 商品分类名称变了
+                // 删除 catNameToTypeID map 缓存中 k=旧的商品分类名称
+                redisTemplate.boundHashOps("catNameToTypeID").delete(oldName);
+                // 将新的商品分类名称缓存
+                redisTemplate.boundHashOps("catNameToTypeID").put(itemCat.getName(), itemCat.getTypeId());
+            } else {
+                // 商品分类名称没变
+                // 更新缓存数据
+                redisTemplate.boundHashOps("catNameToTypeID").put(itemCat.getName(), itemCat.getTypeId());
+            }
+        }
     }
 
 
@@ -94,7 +140,7 @@ public class ItemCatServiceImpl implements ItemCatService {
 
 
     /**
-     * 批量删除商品分类
+     * 批量删除商品分类（删除数据库中的内容，删除缓存中的内容）
      */
     @Override
     public void batchDelete(Long[] ids) throws Exception {
@@ -106,7 +152,20 @@ public class ItemCatServiceImpl implements ItemCatService {
             if (tbItemCats != null && tbItemCats.size() > 0) {
                 throw new Exception();
             } else {
-                tbItemCatDao.deleteById(id);
+
+                // 获取商品分类的名称
+                TbItemCat tb = new TbItemCat();
+                tb.setId(id);
+                List<TbItemCat> result = tbItemCatDao.queryAll(tb);
+                if (result != null && result.size() > 0) {
+                    String name = result.get(0).getName();
+
+                    // 删除数据库中的商品分类
+                    tbItemCatDao.deleteById(id);
+
+                    // 删除 catNameToTypeID map 缓存中 k=旧的商品分类名称
+                    redisTemplate.boundHashOps("catNameToTypeID").delete(name);
+                }
             }
         }
     }
